got rid of warnings on Linux g++8

+++ macro.pfmoi/src/Human-PfMOI.cpp
@@ -263,7 +263,7 @@ void human::queue_bites(){
 
   if(nbite > 0){
     double tnow = tileP->get_tnow();
-    for(size_t i=0; i<nbite; i++){
+    for(int i=0; i<nbite; i++){
       addEvent2Q(e_pfmoi_bite(tnow,this));
     }
   }
